pyproject.toml
[project]
name = "chat-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.13",
    "litellm>=1.73.0",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.0",
    "pypdf2>=3.0.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=4.1.0",
    "structlog>=25.4.0",
    "transformers>=4.52.4",
    "uvicorn>=0.34.3",
]
